Refactor to search-information component
