+++ frontend/src/components/index.js
@@ -1,0 +1,7 @@
+import SearchBar from './search-bar'
+import SearchInformation from './search-information'
+
+export {
+    SearchBar,
+    SearchInformation
+}
